refactor: Expressアプリ定義をapp.jsに分離して起動スクリプトを簡素化
